moved print_memory_diags into utils
